+++ applications/auto_setup/backend/setup_linac_object.py
