+++ applications/auto_setup/setup_gui.py
@@ -23,10 +23,10 @@ from pydm.widgets import PyDMLabel
 from pydm.widgets.analog_indicator import PyDMAnalogIndicator
 from pydm.widgets.display_format import DisplayFormat
 
-from applications.auto_setup.setup_cavity import SetupCavity
-from applications.auto_setup.setup_cryomodule import SetupCryomodule
-from applications.auto_setup.setup_machine import SETUP_MACHINE
-from setup_linac import SetupLinac
+from applications.auto_setup.backend.setup_cavity import SetupCavity
+from applications.auto_setup.backend.setup_cryomodule import SetupCryomodule
+from applications.auto_setup.backend.setup_linac import SetupLinac
+from applications.auto_setup.backend.setup_machine import SETUP_MACHINE
 from utils.sc_linac import linac_utils
