+++ applications/auto_setup/srf_global_setup_launcher.py
@@ -1,7 +1,7 @@
 import argparse
 
-from applications.auto_setup.setup_cryomodule import SetupCryomodule
-from applications.auto_setup.setup_machine import SetupMachine, SETUP_MACHINE
+from applications.auto_setup.backend.setup_cryomodule import SetupCryomodule
+from applications.auto_setup.backend.setup_machine import SetupMachine, SETUP_MACHINE
 from utils.sc_linac.linac_utils import ALL_CRYOMODULES_NO_HL, ALL_CRYOMODULES
