+++ applications/auto_setup/srf_linac_setup_launcher.py
@@ -1,8 +1,8 @@
 import argparse
 from time import sleep
 
-from applications.auto_setup.setup_cryomodule import SetupCryomodule
-from applications.auto_setup.setup_machine import SETUP_MACHINE
+from applications.auto_setup.backend.setup_cryomodule import SetupCryomodule
+from applications.auto_setup.backend.setup_machine import SETUP_MACHINE
 from utils.sc_linac.linac_utils import LINAC_CM_DICT
