+++ tests/applications/auto_setup/test_setup_machine.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from applications.auto_setup.setup_machine import SetupMachine
+from applications.auto_setup.backend.setup_machine import SetupMachine
 
 
 @pytest.fixture
